--- src/main/java/com/woowacourse/pelotonbackend/member/Role.java
@@ -1,5 +1,0 @@
-package com.woowacourse.pelotonbackend.member;
-
-public enum Role {
-    MEMBER, ADMIN
-}
