+++ src/main/java/com/woowacourse/pelotonbackend/member/domain/MemberRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.pelotonbackend.member.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface MemberRepository extends CrudRepository<Member, Long> {
+}
