+++ src/main/java/com/woowacourse/pelotonbackend/member/domain/Role.java
@@ -1,0 +1,5 @@
+package com.woowacourse.pelotonbackend.member.domain;
+
+public enum Role {
+    MEMBER, ADMIN
+}
