+++ src/main/java/com/woowacourse/pelotonbackend/mission/domain/MissionRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.pelotonbackend.mission.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface MissionRepository extends CrudRepository<Mission, Long> {
+}
