+++ src/main/java/com/woowacourse/pelotonbackend/missioncertification/domain/MissionCertification.java
@@ -1,17 +1,18 @@
-package com.woowacourse.pelotonbackend.member;
+package com.woowacourse.pelotonbackend.missioncertification.domain;
 
 import java.time.LocalDateTime;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Embedded;
 
-import com.woowacourse.pelotonbackend.vo.Cash;
+import com.woowacourse.pelotonbackend.mission.domain.Mission;
+import com.woowacourse.pelotonbackend.rider.domain.Rider;
+import com.woowacourse.pelotonbackend.vo.ImageUrl;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -23,22 +24,23 @@ import lombok.With;
 @Builder
 @EqualsAndHashCode(of = "id")
 @Getter
-public class Member {
+public class MissionCertification {
     @Id @With(value = AccessLevel.PACKAGE)
     private final Long id;
 
-    @NotBlank
-    private final String name;
+    @NotNull
+    private final MissionCertificationStatus status;
 
-    @NotBlank
-    @Email
-    private final String email;
+    private final String description;
 
     @Embedded.Empty
-    private final Cash cash;
+    private final ImageUrl image;
 
     @NotNull
-    private final Role role;
+    private final AggregateReference<Rider, @NotNull Long> riderId;
+
+    @NotNull
+    private final AggregateReference<Mission, @NotNull Long> missionId;
 
     @CreatedDate
     private LocalDateTime createdAt;
