+++ src/main/java/com/woowacourse/pelotonbackend/missioncertification/domain/MissionCertificationRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.pelotonbackend.missioncertification.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface MissionCertificationRepository extends CrudRepository<MissionCertification, Long> {
+}
