+++ src/main/java/com/woowacourse/pelotonbackend/missioncertificationreport/domain/MissionCertificationReport.java
@@ -1,7 +1,8 @@
-package com.woowacourse.pelotonbackend.rider;
+package com.woowacourse.pelotonbackend.missioncertificationreport.domain;
 
 import java.time.LocalDateTime;
 
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 import org.springframework.data.annotation.CreatedDate;
@@ -9,8 +10,8 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 
-import com.woowacourse.pelotonbackend.member.Member;
-import com.woowacourse.pelotonbackend.race.Race;
+import com.woowacourse.pelotonbackend.member.domain.Member;
+import com.woowacourse.pelotonbackend.missioncertification.domain.MissionCertification;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,15 +23,21 @@ import lombok.With;
 @Builder
 @EqualsAndHashCode(of = "id")
 @Getter
-public class Rider {
+public class MissionCertificationReport {
     @Id @With(value = AccessLevel.PACKAGE)
     private final Long id;
 
     @NotNull
-    private final AggregateReference<Member, @NotNull Long> memberId;
+    private final ReportType reportType;
+
+    @NotBlank
+    private final String description;
 
     @NotNull
-    private final AggregateReference<Race, @NotNull Long> raceId;
+    private final AggregateReference<MissionCertification, @NotNull Long> missionCertificationId;
+
+    @NotNull
+    private final AggregateReference<Member, @NotNull Long> memberId;
 
     @CreatedDate
     private LocalDateTime createdAt;
