+++ src/main/java/com/woowacourse/pelotonbackend/missioncertificationreport/domain/MissionCertificationReportRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.pelotonbackend.missioncertificationreport.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface MissionCertificationReportRepository extends CrudRepository<MissionCertificationReport, Long> {
+}
