+++ src/main/java/com/woowacourse/pelotonbackend/race/domain/RaceCategory.java
@@ -1,0 +1,5 @@
+package com.woowacourse.pelotonbackend.race.domain;
+
+public enum RaceCategory {
+    TIME
+}
