+++ src/main/java/com/woowacourse/pelotonbackend/race/domain/RaceRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.pelotonbackend.race.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface RaceRepository extends CrudRepository<Race, Long> {
+}
