+++ src/main/java/com/woowacourse/pelotonbackend/rider/domain/RiderRepository.java
@@ -1,0 +1,6 @@
+package com.woowacourse.pelotonbackend.rider.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface RiderRepository extends CrudRepository<Rider, Long> {
+}
